sentiment_app/requirements.txt
streamlit
langchain==0.2.17
langchain-community==0.2.7
groq
python-dotenv
requests==2.32.4